Rust version up. fix iter. README.md - update

+++ emcc_c/sumInt.c
@@ -1,13 +1,13 @@
-int sumInt(int *array, int n) {
-  int s = 0;
+long long sumInt(int *array, int n) {
+  long long s = 0;
   for (int i = 0; i < n; i++) {
     s += array[i];
   }
   return s;
 }
 
-int inlineSumInt(int *array, int n, int iter) {
-  int s = 0;
+long long inlineSumInt(int *array, int n, int iter) {
+  long long s = 0;
   for (int i = 0 ; i < iter; i++) {
     s += sumInt(array, n);
   }
